add func to logged out users managing user settings

+++ src/utils/blitz-utils.ts
@@ -67,12 +67,10 @@ export const formatZodError = (err: any) => {
   });
 };
 
-const EMAIL_VERIFY_LINK_IN_HOURS = 1;
-
-const createToken = async ({ userId, userEmail, tokenType }) => {
+const createToken = async ({ userId, userEmail, tokenType, expiryHours = 4 }) => {
   const token = generateToken();
   const hashedToken = hash256(token);
-  const expiresAt = addHours(new Date(), EMAIL_VERIFY_LINK_IN_HOURS);
+  const expiresAt = addHours(new Date(), expiryHours);
 
   await db.token.create({
     data: {
@@ -90,19 +88,27 @@ export const regenerateToken = async ({
   userId,
   userEmail,
   tokenType,
+  expiryHours = 4,
+  deleteExisting = true,
 }: {
   userId: string;
   userEmail: string;
   tokenType: TokenType;
+  expiryHours?: number;
+  deleteExisting?: boolean;
 }): Promise<string> => {
-  await db.token.deleteMany({
-    where: { type: tokenType, userId },
-  });
+
+  if(deleteExisting) {
+    await db.token.deleteMany({
+      where: { type: tokenType, userId },
+    });
+  }
 
   const token = await createToken({
     userId,
     userEmail,
     tokenType,
+    expiryHours
   });
 
   return token;
